rafactor: LottoReport, LottoStore 통합
