+++ src/main/java/lotto/model/LottoStore.java
@@ -1,5 +1,9 @@
 package lotto.model;
 
+import common.model.Number;
+
+import java.util.Map;
+
 public class LottoStore {
 
     private static final int LOTTO_PRICE = 1000;
@@ -12,7 +16,14 @@ public class LottoStore {
         return new Lotto(count);
     }
 
-    public static LottoReport report(Lotto lotto, LottoWinner winner) {
-        return new LottoReport(lotto, winner);
+    public static float calculateRateOfRevenue(Map<LottoRank, Number> lottoRankNumbers, int purchaseAmount) {
+        float revenue = 0;
+
+        for (LottoRank rank : LottoRank.valuesWithoutMiss()) {
+            Number count = lottoRankNumbers.getOrDefault(rank, new Number());
+            revenue += rank.getAmount() * count.getNumber();
+        }
+
+        return Math.round(revenue / purchaseAmount * 100) / 100.F;
     }
 }
